+++ src/app/apps/dashboards/views.py
@@ -1,0 +1,7 @@
+from django.views.generic import TemplateView
+
+
+class DashboardView(TemplateView):
+    """Dummy dashboard view"""
+
+    template_name = "dashboards/base.html"