+++ src/app/apps/lti/apps.py
@@ -1,0 +1,10 @@
+"""The LTI app."""
+
+from django.apps import AppConfig
+
+
+class LTIConfig(AppConfig):
+    """Configuration for the LTI app."""
+
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "apps.lti"