+++ src/app/docker/files/usr/local/etc/gunicorn/warren.py
@@ -1,0 +1,18 @@
+"""Gunicorn configuration file for warren."""
+
+# Gunicorn-django settings
+bind = ["0.0.0.0:8000"]
+name = "warren"
+python_path = "/app"
+
+# Run
+graceful_timeout = 90
+timeout = 90
+workers = 3
+
+# Logging
+# Using '-' for the access log file makes gunicorn log accesses to stdout
+accesslog = "-"
+# Using '-' for the error log file makes gunicorn log errors to stderr
+errorlog = "-"
+loglevel = "info"